+++ include/boost/mysql/impl/internal/sansio/connection_state_data.hpp
@@ -14,7 +14,7 @@
 #include <boost/mysql/metadata_mode.hpp>
 
 #include <boost/mysql/impl/internal/protocol/capabilities.hpp>
-#include <boost/mysql/impl/internal/protocol/constants.hpp>
+#include <boost/mysql/impl/internal/protocol/db_flavor.hpp>
 #include <boost/mysql/impl/internal/sansio/message_reader.hpp>
 #include <boost/mysql/impl/internal/sansio/message_writer.hpp>
