+++ include/boost/mysql/impl/internal/sansio/message_reader.hpp
@@ -11,8 +11,8 @@
 #include <boost/mysql/client_errc.hpp>
 #include <boost/mysql/error_code.hpp>
 
-#include <boost/mysql/impl/internal/protocol/constants.hpp>
 #include <boost/mysql/impl/internal/protocol/deserialization.hpp>
+#include <boost/mysql/impl/internal/protocol/frame_header.hpp>
 #include <boost/mysql/impl/internal/sansio/read_buffer.hpp>
 
 #include <boost/asio/coroutine.hpp>
