+++ include/boost/mysql/impl/internal/sansio/quit_connection.hpp
@@ -13,13 +13,12 @@
 
 #include <boost/mysql/detail/algo_params.hpp>
 
+#include <boost/mysql/impl/internal/protocol/serialization.hpp>
 #include <boost/mysql/impl/internal/sansio/connection_state_data.hpp>
 #include <boost/mysql/impl/internal/sansio/next_action.hpp>
 #include <boost/mysql/impl/internal/sansio/sansio_algorithm.hpp>
 
 #include <boost/asio/coroutine.hpp>
-
-#include <cstddef>
 
 namespace boost {
 namespace mysql {
