+++ test/fuzzing/fuzz_prepare_stmt_response.cpp
@@ -7,7 +7,7 @@
 
 #include <boost/mysql/diagnostics.hpp>
 
-#include <boost/mysql/impl/internal/protocol/constants.hpp>
+#include <boost/mysql/impl/internal/protocol/db_flavor.hpp>
 #include <boost/mysql/impl/internal/protocol/deserialization.hpp>
 
 using namespace boost::mysql::detail;
