+++ test/fuzzing/fuzz_server_hello.cpp
@@ -7,7 +7,7 @@
 
 #include <boost/mysql/diagnostics.hpp>
 
-#include <boost/mysql/impl/internal/protocol/constants.hpp>
+#include <boost/mysql/impl/internal/protocol/db_flavor.hpp>
 #include <boost/mysql/impl/internal/protocol/deserialization.hpp>
 
 using namespace boost::mysql::detail;
